Subscribe to logs for new pipelines
These logs are shipped to the OST endpoint.
These logs are created implicitly by CodePipeline so the name cannot be
inferred from the terraform resource. They have to be referenced by
name.

To send future logs to OST from this shared services account, add the
log group name to this list.