Create Logging Heartbeat
In order to keep development, pre-production and production the same for
the logging service, we want to emulate sending logs from Shared
services to each one of these accounts.

You can only have a maximum of 2 subsrciptions on a log group so cannot
use cloudtrail to send logs.

Create a dedicated log group with messages going into it every 2
seconds.