Amend shebang
Running script files through /usr/bin/env has the advantage of automatically searching for the default version of the interpreter in the current environment. This way, we don’t need to search for it in a specific location in the system, as the paths may be different in other systems.